+++ emBODY/eBcode/arch-arm/embobj/plus/mc/EOaxisController.h
@@ -90,7 +90,7 @@ extern int32_t eo_axisController_GetAxisPos (EOaxisController *o);
 extern int32_t eo_axisController_GetAxisVel (EOaxisController *o);
 extern void eo_axisController_SetEncPos(EOaxisController *o, int32_t pos); 
 extern void eo_axisController_SetEncVel(EOaxisController *o, int32_t vel);
-extern void eo_axisController_SetTorque(EOaxisController *o, int16_t trq);
+extern void eo_axisController_SetTorque(EOaxisController *o, int32_t trq);
 
 //extern int32_t eo_axisController_GetVelocity(EOaxisController *o);
